test: split media runtime coverage
