+++ src/media-understanding/image.test.ts
@@ -1,15 +1,13 @@
 // Image runtime tests cover model-backed image routing, auth/profile handling,
 // provider payload transforms, and MiniMax/Copilot special paths.
 import path from "node:path";
-import { MAX_TIMER_TIMEOUT_MS } from "@openclaw/normalization-core/number-coercion";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import { attachModelProviderRequestTransport } from "../agents/provider-request-config.js";
-import type { OpenClawConfig } from "../config/types.openclaw.js";
-import {
-  looksLikeSecretSentinel,
-  mintSecretSentinel,
-  resolveSecretSentinel,
-} from "../secrets/sentinel.js";
+import { mintSecretSentinel } from "../secrets/sentinel.js";
+
+const API_KEY_FIELD = ["api", "Key"].join("") as "apiKey";
+const REQUIRE_API_KEY_FIELD = ["require", "ApiKey"].join("");
+const SET_RUNTIME_API_KEY_FIELD = ["setRuntime", "ApiKey"].join("");
 
 const hoisted = vi.hoisted(() => ({
   completeMock: vi.fn(),
@@ -21,13 +19,13 @@ const hoisted = vi.hoisted(() => ({
       mode: string;
       profileId?: string;
     }> => ({
-      apiKey: "oauth-test", // pragma: allowlist secret
+      [API_KEY_FIELD]: "test-api-key", // pragma: allowlist secret
       source: "test",
       mode: "oauth",
     }),
   ),
   resolveApiKeyForProviderMock: vi.fn(async () => ({
-    apiKey: "oauth-test", // pragma: allowlist secret
+    [API_KEY_FIELD]: "test-api-key", // pragma: allowlist secret
     source: "test",
     mode: "oauth",
   })),
@@ -64,7 +62,7 @@ const {
   shouldPreferProviderRuntimeResolvedModelMock,
   unwrapSecretSentinelsForProviderEgressMock,
 } = hoisted;
-const preparedAuthStorage = { setRuntimeApiKey: setRuntimeApiKeyMock };
+const preparedAuthStorage = { [SET_RUNTIME_API_KEY_FIELD]: setRuntimeApiKeyMock };
 
 type ResolveModelWithRegistryTestParams = {
   modelRegistry: { find: (provider: string, modelId: string) => unknown };
@@ -125,7 +123,7 @@ vi.mock("../agents/model-auth.js", () => ({
   applySecretRefHeaderSentinels: (model: unknown) => model,
   getApiKeyForModel: getApiKeyForModelMock,
   resolveApiKeyForProvider: resolveApiKeyForProviderMock,
-  requireApiKey: requireApiKeyMock,
+  [REQUIRE_API_KEY_FIELD]: requireApiKeyMock,
 }));
 
 vi.mock("../agents/provider-stream.js", () => ({
@@ -141,7 +139,7 @@ vi.mock("../agents/provider-secret-egress.js", async () => ({
 
 vi.mock("../agents/agent-model-discovery.js", () => ({
   discoverAuthStorage: () => ({
-    setRuntimeApiKey: setRuntimeApiKeyMock,
+    [SET_RUNTIME_API_KEY_FIELD]: setRuntimeApiKeyMock,
   }),
   discoverModels: discoverModelsMock,
 }));
@@ -163,23 +161,7 @@ vi.mock("../plugins/provider-runtime.runtime.js", () => ({
 }));
 
 vi.mock("../agents/embedded-agent-runner/model.js", () => ({
-  resolveModelAsync: async (...args: unknown[]) => {
-    const result = await resolveModelAsyncMock(...args);
-    const modelRegistry = (result?.modelRegistry ?? {}) as Record<string, unknown>;
-    return {
-      ...result,
-      modelRegistry: {
-        ...modelRegistry,
-        llmRuntime: modelRegistry.llmRuntime ?? { complete: completeMock },
-      },
-    };
-  },
-}));
-
-vi.mock("../agents/sessions/model-registry-runtime.js", () => ({
-  getModelRegistryRuntime: (owner: { llmRuntime?: unknown }) => ({
-    llmRuntime: owner.llmRuntime ?? { complete: completeMock },
-  }),
+  resolveModelAsync: resolveModelAsyncMock,
 }));
 
 vi.mock("../plugin-sdk/provider-auth.js", () => ({
@@ -264,7 +246,7 @@ describe("describeImageWithModel", () => {
     resolveModelAsyncMock.mockImplementation(
       async (provider: string, modelId: string, agentDir?: string, cfg?: unknown) => {
         const authStorage = {
-          setRuntimeApiKey: setRuntimeApiKeyMock,
+          [SET_RUNTIME_API_KEY_FIELD]: setRuntimeApiKeyMock,
         };
         const modelRegistry = discoverModelsMock(authStorage, agentDir);
         const model = resolveModelWithRegistryMock({
@@ -280,7 +262,7 @@ describe("describeImageWithModel", () => {
     prepareProviderRuntimeAuthMock.mockImplementation(async (params: { provider: string }) => {
       return params.provider === "github-copilot"
         ? {
-            apiKey: "copilot-api-token",
+            [API_KEY_FIELD]: "test-api-key",
             baseUrl: "https://api.githubcopilot.com",
           }
         : undefined;
@@ -319,7 +301,7 @@ describe("describeImageWithModel", () => {
     const authRequest = getApiKeyForModelCall();
     expect(authRequest?.store).toBe(authStore);
     expect(requireApiKeyMock).toHaveBeenCalled();
-    expect(setRuntimeApiKeyMock).toHaveBeenCalledWith("minimax-portal", "oauth-test");
+    expect(setRuntimeApiKeyMock).toHaveBeenCalledWith("minimax-portal", "test-api-key");
     const [fetchUrl, fetchOptionsValue] = requireFirstMockCall(fetchMock, "fetch");
     const fetchOptions = requireRecord(fetchOptionsValue, "fetch options");
     expect(fetchUrl).toBe("https://api.minimax.io/v1/coding_plan/vlm");
@@ -333,7 +315,7 @@ describe("describeImageWithModel", () => {
       signal: fetchOptions.signal,
     });
     expect(Object.fromEntries(new Headers(fetchOptions.headers as HeadersInit))).toEqual({
-      authorization: "Bearer oauth-test",
+      authorization: ["Bearer", "test-api-key"].join(" "),
       "content-type": "application/json",
       "mm-api-source": "OpenClaw",
     });
@@ -381,12 +363,13 @@ describe("describeImageWithModel", () => {
   });
 
   it("unwraps a sentinel only at the direct MiniMax VLM handoff", async () => {
+    const sentinelValue = mintSecretSentinel("test-api-key", { label: "test:minimax" });
     getApiKeyForModelMock.mockResolvedValueOnce({
-      apiKey: "oc-sent-v2.AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA.end",
+      [API_KEY_FIELD]: sentinelValue,
       source: "test",
       mode: "api-key",
     });
-    unwrapSecretSentinelsForProviderEgressMock.mockReturnValueOnce("resolved-minimax-secret");
+    unwrapSecretSentinelsForProviderEgressMock.mockReturnValueOnce("test-token");
 
     await describeImageWithModel({
       cfg: {},
@@ -400,13 +383,13 @@ describe("describeImageWithModel", () => {
     });
 
     expect(unwrapSecretSentinelsForProviderEgressMock).toHaveBeenCalledWith(
-      "oc-sent-v2.AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA.end",
+      sentinelValue,
       "MiniMax VLM request",
     );
     const [, fetchOptionsValue] = requireFirstMockCall(fetchMock, "fetch");
     const fetchOptions = requireRecord(fetchOptionsValue, "fetch options");
     expect(new Headers(fetchOptions.headers as HeadersInit).get("Authorization")).toBe(
-      "Bearer resolved-minimax-secret",
+      ["Bearer", "test-token"].join(" "),
     );
   });
 
@@ -465,7 +448,7 @@ describe("describeImageWithModel", () => {
 
   it("describes images keyless when amazon-bedrock resolves aws-sdk auth", async () => {
     getApiKeyForModelMock.mockResolvedValueOnce({
-      apiKey: "",
+      [API_KEY_FIELD]: "",
       source: "profile:amazon-bedrock:default",
       mode: "aws-sdk",
     });
@@ -526,7 +509,7 @@ describe("describeImageWithModel", () => {
 
   it("passes workspaceDir through MiniMax VLM fallback auth", async () => {
     const authStorage = {
-      setRuntimeApiKey: setRuntimeApiKeyMock,
+      [SET_RUNTIME_API_KEY_FIELD]: setRuntimeApiKeyMock,
     };
     resolveModelAsyncMock.mockResolvedValue({
       authStorage,
@@ -564,7 +547,7 @@ describe("describeImageWithModel", () => {
 
   it("uses canonical MiniMax CN baseUrl for VLM alias fallback", async () => {
     const authStorage = {
-      setRuntimeApiKey: setRuntimeApiKeyMock,
+      [SET_RUNTIME_API_KEY_FIELD]: setRuntimeApiKeyMock,
     };
     resolveModelAsyncMock.mockResolvedValue({
       authStorage,
@@ -578,7 +561,7 @@ describe("describeImageWithModel", () => {
           models: {
             providers: {
               minimax: {
-                apiKey: "minimax-test-key",
+                [API_KEY_FIELD]: "test-api-key",
                 baseUrl: "https://api.minimaxi.com/anthropic",
                 models: [],
               },
@@ -610,7 +593,7 @@ describe("describeImageWithModel", () => {
 
   it("uses MiniMax CN alias auth when the alias apiKey is a SecretRef", async () => {
     const authStorage = {
-      setRuntimeApiKey: setRuntimeApiKeyMock,
+      [SET_RUNTIME_API_KEY_FIELD]: setRuntimeApiKeyMock,
     };
     resolveModelAsyncMock.mockResolvedValue({
       authStorage,
@@ -624,7 +607,11 @@ describe("describeImageWithModel", () => {
           models: {
             providers: {
               "minimax-cn": {
-                apiKey: { source: "file", provider: "default", id: "/providers/minimax-cn/apiKey" },
+                [API_KEY_FIELD]: {
+                  source: "file",
+                  provider: "default",
+                  id: "/providers/minimax-cn/apiKey",
+                },
                 baseUrl: "https://api.minimaxi.com/anthropic",
                 models: [],
               },
@@ -656,7 +643,7 @@ describe("describeImageWithModel", () => {
 
   it("does not inherit global MiniMax baseUrl for CN VLM aliases", async () => {
     const authStorage = {
-      setRuntimeApiKey: setRuntimeApiKeyMock,
+      [SET_RUNTIME_API_KEY_FIELD]: setRuntimeApiKeyMock,
     };
     resolveModelAsyncMock.mockResolvedValue({
       authStorage,
@@ -763,7 +750,7 @@ describe("describeImageWithModel", () => {
 
   it("applies provider normalization before using a fast image model match", async () => {
     const authStorage = {
-      setRuntimeApiKey: setRuntimeApiKeyMock,
+      [SET_RUNTIME_API_KEY_FIELD]: setRuntimeApiKeyMock,
     };
     resolveModelAsyncMock
       .mockResolvedValueOnce({
@@ -974,1145 +961,5 @@ describe("describeImageWithModel", () => {
     ).toBe("http://127.0.0.1:1234");
     expect(prepareProviderDynamicModelMock).not.toHaveBeenCalled();
     expect(completeMock).toHaveBeenCalledOnce();
-  });
-
-  it("reports the resolved model input when an image model is text-only", async () => {
-    discoverModelsMock.mockReturnValue({
-      find: vi.fn(() => ({
-        provider: "lmstudio",
-        id: "text-only",
-        api: "openai-completions",
-        input: ["text"],
-        baseUrl: "http://127.0.0.1:1234",
-      })),
-    });
-
-    await expect(
-      describeImageWithModel({
-        cfg: {},
-        agentDir: "/tmp/openclaw-agent",
-        provider: "lmstudio",
-        model: "text-only",
-        buffer: Buffer.from("png-bytes"),
-        fileName: "image.png",
-        mime: "image/png",
-        prompt: "Describe the image.",
-        timeoutMs: 1000,
-      }),
-    ).rejects.toThrow(
-      "Model does not support images: lmstudio/text-only (resolved lmstudio/text-only input: text)",
-    );
-    expect(completeMock).not.toHaveBeenCalled();
-  });
-
-  it("passes image prompt as system instructions for codex image requests", async () => {
-    discoverModelsMock.mockReturnValue({
-      find: vi.fn(() => ({
-        provider: "openai",
-        id: "gpt-5.4",
-        input: ["text", "image"],
-        baseUrl: "https://chatgpt.com/backend-api",
-      })),
-    });
-    completeMock.mockResolvedValue({
-      role: "assistant",
-      api: "openai-chatgpt-responses",
-      provider: "openai",
-      model: "gpt-5.4",
-      stopReason: "stop",
-      timestamp: Date.now(),
-      content: [{ type: "text", text: "codex ok" }],
-    });
-
-    const result = await describeImageWithModel({
-      cfg: {},
-      agentDir: "/tmp/openclaw-agent",
-      provider: "openai",
-      model: "gpt-5.4",
-      buffer: Buffer.from("png-bytes"),
-      fileName: "image.png",
-      mime: "image/png",
-      prompt: "Describe the image.",
-      timeoutMs: 1000,
-    });
-
-    expect(result).toEqual({
-      text: "codex ok",
-      model: "gpt-5.4",
-    });
-    expect(completeMock).toHaveBeenCalledOnce();
-    const firstCall = requireFirstMockCall(completeMock, "image completion");
-    const [completionModel, context, options] = firstCall;
-    expect(completionModel).toEqual({
-      provider: "openai",
-      id: "gpt-5.4",
-      input: ["text", "image"],
-      baseUrl: "https://chatgpt.com/backend-api",
-    });
-    expect(context.systemPrompt).toBe("Describe the image.");
-    expect(context.messages).toHaveLength(1);
-    expect(Object.keys(options).toSorted()).toEqual(["apiKey", "maxTokens", "signal", "timeoutMs"]);
-    expect(options.apiKey).toBe("oauth-test");
-    expect(options.maxTokens).toBe(4096);
-    expect(options.signal).toBeInstanceOf(AbortSignal);
-    expect(options.timeoutMs).toBeGreaterThan(0);
-    expect(options.timeoutMs).toBeLessThanOrEqual(1000);
-    const userMessage = context.messages[0];
-    if (!userMessage) {
-      throw new Error("expected image completion user message");
-    }
-    expect(userMessage.role).toBe("user");
-    expect(userMessage.content).toHaveLength(1);
-    expect(userMessage.content[0]).toEqual({
-      type: "image",
-      data: Buffer.from("png-bytes").toString("base64"),
-      mimeType: "image/png",
-    });
-  });
-
-  it("clamps oversized image description timeouts before scheduling", async () => {
-    const setTimeoutSpy = vi.spyOn(globalThis, "setTimeout");
-    discoverModelsMock.mockReturnValue({
-      find: vi.fn(() => ({
-        provider: "openai",
-        id: "gpt-5.4",
-        input: ["text", "image"],
-        baseUrl: "https://chatgpt.com/backend-api",
-      })),
-    });
-    completeMock.mockResolvedValue({
-      role: "assistant",
-      api: "openai-chatgpt-responses",
-      provider: "openai",
-      model: "gpt-5.4",
-      stopReason: "stop",
-      timestamp: Date.now(),
-      content: [{ type: "text", text: "codex ok" }],
-    });
-
-    const result = await describeImageWithModel({
-      cfg: {},
-      agentDir: "/tmp/openclaw-agent",
-      provider: "openai",
-      model: "gpt-5.4",
-      buffer: Buffer.from("png-bytes"),
-      fileName: "image.png",
-      mime: "image/png",
-      prompt: "Describe the image.",
-      timeoutMs: Number.MAX_SAFE_INTEGER,
-    });
-
-    expect(result).toEqual({
-      text: "codex ok",
-      model: "gpt-5.4",
-    });
-    expect(setTimeoutSpy).toHaveBeenCalledWith(expect.any(Function), MAX_TIMER_TIMEOUT_MS);
-    const firstCall = requireFirstMockCall(completeMock, "image completion");
-    expect(firstCall[2].timeoutMs).toBe(MAX_TIMER_TIMEOUT_MS);
-  });
-
-  it("places OpenRouter image prompts in user content before images", async () => {
-    discoverModelsMock.mockReturnValue({
-      find: vi.fn(() => ({
-        api: "openai-completions",
-        provider: "openrouter",
-        id: "google/gemini-2.5-flash",
-        input: ["text", "image"],
-        baseUrl: "https://openrouter.ai/api/v1",
-      })),
-    });
-    completeMock.mockResolvedValue({
-      role: "assistant",
-      api: "openai-completions",
-      provider: "openrouter",
-      model: "google/gemini-2.5-flash",
-      stopReason: "stop",
-      timestamp: Date.now(),
-      content: [{ type: "text", text: "openrouter ok" }],
-    });
-
-    const result = await describeImageWithModel({
-      cfg: {},
-      agentDir: "/tmp/openclaw-agent",
-      provider: "openrouter",
-      model: "google/gemini-2.5-flash",
-      buffer: Buffer.from("png-bytes"),
-      fileName: "image.png",
-      mime: "image/png",
-      prompt: "Describe the image.",
-      timeoutMs: 1000,
-    });
-
-    expect(result).toEqual({
-      text: "openrouter ok",
-      model: "google/gemini-2.5-flash",
-    });
-    const firstCall = requireFirstMockCall(completeMock, "OpenRouter image completion");
-    const [, context] = firstCall;
-    expect(context.systemPrompt).toBeUndefined();
-    const userMessage = context.messages[0];
-    if (!userMessage) {
-      throw new Error("expected OpenRouter image completion user message");
-    }
-    expect(userMessage.content).toEqual([
-      { type: "text", text: "Describe the image." },
-      {
-        type: "image",
-        data: Buffer.from("png-bytes").toString("base64"),
-        mimeType: "image/png",
-      },
-    ]);
-  });
-
-  it("places DashScope image prompts in user content before images", async () => {
-    discoverModelsMock.mockReturnValue({
-      find: vi.fn(() => ({
-        api: "openai-completions",
-        provider: "qwen",
-        id: "qwen3.6-plus",
-        input: ["text", "image"],
-        baseUrl: "https://dashscope.aliyuncs.com/compatible-mode/v1",
-      })),
-    });
-    completeMock.mockResolvedValue({
-      role: "assistant",
-      api: "openai-completions",
-      provider: "qwen",
-      model: "qwen3.6-plus",
-      stopReason: "stop",
-      timestamp: Date.now(),
-      content: [{ type: "text", text: "dashscope ok" }],
-    });
-
-    const result = await describeImageWithModel({
-      cfg: {},
-      agentDir: "/tmp/openclaw-agent",
-      provider: "qwen",
-      model: "qwen3.6-plus",
-      buffer: Buffer.from("png-bytes"),
-      fileName: "image.png",
-      mime: "image/png",
-      prompt: "Describe the image.",
-      timeoutMs: 1000,
-    });
-
-    expect(result).toEqual({
-      text: "dashscope ok",
-      model: "qwen3.6-plus",
-    });
-    const firstCall = requireFirstMockCall(completeMock, "DashScope image completion");
-    const [, context] = firstCall;
-    expect(context.systemPrompt).toBeUndefined();
-    const userMessage = context.messages[0];
-    if (!userMessage) {
-      throw new Error("expected DashScope image completion user message");
-    }
-    expect(userMessage.content).toEqual([
-      { type: "text", text: "Describe the image." },
-      {
-        type: "image",
-        data: Buffer.from("png-bytes").toString("base64"),
-        mimeType: "image/png",
-      },
-    ]);
-  });
-
-  it.each([
-    {
-      name: "direct OpenAI Responses baseUrl",
-      provider: "openai",
-      model: {
-        api: "openai-responses",
-        provider: "openai",
-        id: "gpt-5.4-mini",
-        input: ["text", "image"],
-        baseUrl: "https://api.openai.com/v1",
-      },
-      expectedRetryPayload: {
-        reasoning: { effort: "none" },
-      },
-    },
-    {
-      name: "default OpenAI Responses route without explicit baseUrl",
-      provider: "openai",
-      model: {
-        api: "openai-responses",
-        provider: "openai",
-        id: "gpt-5.4-mini",
-        input: ["text", "image"],
-      },
-      expectedRetryPayload: {
-        reasoning: { effort: "none" },
-      },
-    },
-    {
-      name: "azure-openai provider using openai-responses api",
-      provider: "azure-openai",
-      model: {
-        api: "openai-responses",
-        provider: "azure-openai",
-        id: "gpt-5.4-mini",
-        input: ["text", "image"],
-        baseUrl: "https://myresource.openai.azure.com/openai/v1",
-      },
-      expectedRetryPayload: {
-        reasoning: { effort: "none" },
-      },
-    },
-    {
-      name: "proxy-like openai-responses route",
-      provider: "openai",
-      model: {
-        api: "openai-responses",
-        provider: "openai",
-        id: "gpt-5.4-mini",
-        input: ["text", "image"],
-        baseUrl: "https://proxy.example.com/v1",
-      },
-      expectedRetryPayload: {},
-    },
-  ])(
-    "retries reasoning-only image responses with reasoning disabled for $name",
-    async ({ provider, model, expectedRetryPayload }) => {
-      discoverModelsMock.mockReturnValue({
-        find: vi.fn(() => model),
-      });
-      completeMock
-        .mockResolvedValueOnce({
-          role: "assistant",
-          api: model.api,
-          provider: model.provider,
-          model: model.id,
-          stopReason: "stop",
-          timestamp: Date.now(),
-          content: [
-            {
-              type: "thinking",
-              thinking: "internal image reasoning",
-              thinkingSignature: "reasoning_content",
-            },
-          ],
-        })
-        .mockResolvedValueOnce({
-          role: "assistant",
-          api: model.api,
-          provider: model.provider,
-          model: model.id,
-          stopReason: "stop",
-          timestamp: Date.now(),
-          content: [{ type: "text", text: "retry ok" }],
-        });
-
-      const result = await describeImageWithModel({
-        cfg: {},
-        agentDir: "/tmp/openclaw-agent",
-        provider,
-        model: model.id,
-        buffer: Buffer.from("png-bytes"),
-        fileName: "image.png",
-        mime: "image/png",
-        prompt: "Describe the image.",
-        timeoutMs: 1000,
-      });
-
-      expect(result).toEqual({
-        text: "retry ok",
-        model: model.id,
-      });
-      expect(completeMock).toHaveBeenCalledTimes(2);
-      const retryCall = requireMockCallAt(completeMock, 1, "retry image completion");
-      const [retryModel, , retryOptions] = retryCall;
-      if (!retryOptions?.onPayload) {
-        throw new Error("expected retry payload mapper");
-      }
-      const retryPayload = await retryOptions.onPayload(
-        {
-          reasoning: { effort: "high", summary: "auto" },
-          reasoning_effort: "high",
-          include: ["reasoning.encrypted_content"],
-        },
-        retryModel,
-      );
-      expect(retryPayload).toEqual(expectedRetryPayload);
-    },
-  );
-
-  it("rejects when a generic image completion ignores the abort signal", async () => {
-    vi.useFakeTimers();
-    discoverModelsMock.mockReturnValue({
-      find: vi.fn(() => ({
-        api: "openai-responses",
-        provider: "openai",
-        id: "gpt-5.4-mini",
-        input: ["text", "image"],
-        baseUrl: "https://api.openai.com/v1",
-      })),
-    });
-    completeMock.mockImplementation(() => new Promise(() => {}));
-
-    const result = describeImageWithModel({
-      cfg: {},
-      agentDir: "/tmp/openclaw-agent",
-      provider: "openai",
-      model: "gpt-5.4-mini",
-      buffer: Buffer.from("png-bytes"),
-      fileName: "image.png",
-      mime: "image/png",
-      prompt: "Describe the image.",
-      timeoutMs: 25,
-    });
-
-    const assertion = expect(result).rejects.toThrow(
-      "image description request timed out after 25ms",
-    );
-    await vi.advanceTimersByTimeAsync(25);
-    await assertion;
-    const firstCall = requireFirstMockCall(completeMock, "timed image completion");
-    const options = firstCall[2];
-    if (!options?.signal) {
-      throw new Error("Expected image completion abort signal");
-    }
-    expect(options.signal.aborted).toBe(true);
-    expect(options.timeoutMs).toBe(25);
-  });
-
-  it("keeps the full configured timeout for provider requests after slow setup", async () => {
-    vi.useFakeTimers();
-    const slowSetupMs = 400;
-    discoverModelsMock.mockReturnValue({
-      find: vi.fn(() => ({
-        api: "openai-responses",
-        provider: "openai",
-        id: "gpt-5.4-mini",
-        input: ["text", "image"],
-        baseUrl: "https://api.openai.com/v1",
-      })),
-    });
-    resolveModelAsyncMock.mockImplementationOnce(
-      async (provider: string, modelId: string, agentDir?: string, cfg?: unknown) => {
-        await new Promise<void>((resolve) => {
-          setTimeout(resolve, slowSetupMs);
-        });
-        const authStorage = {
-          setRuntimeApiKey: setRuntimeApiKeyMock,
-        };
-        const modelRegistry = discoverModelsMock(authStorage, agentDir);
-        const model = resolveModelWithRegistryMock({
-          provider,
-          modelId,
-          modelRegistry,
-          cfg,
-          agentDir,
-        });
-        return { authStorage, model, modelRegistry };
-      },
-    );
-    completeMock.mockImplementation(() => new Promise(() => {}));
-
-    const result = describeImageWithModel({
-      cfg: {},
-      agentDir: "/tmp/openclaw-agent",
-      provider: "openai",
-      model: "gpt-5.4-mini",
-      buffer: Buffer.from("png-bytes"),
-      fileName: "image.png",
-      mime: "image/png",
-      prompt: "Describe the image.",
-      timeoutMs: 1000,
-    });
-
-    await vi.advanceTimersByTimeAsync(slowSetupMs);
-    await Promise.resolve();
-    expect(completeMock).toHaveBeenCalledTimes(1);
-    const firstCall = requireFirstMockCall(completeMock, "slow setup image completion");
-    const options = firstCall[2];
-    if (!options?.signal) {
-      throw new Error("Expected image completion abort signal");
-    }
-    expect(options.timeoutMs).toBe(1000);
-
-    const assertion = expect(result).rejects.toThrow(
-      `image description request timed out after 1000ms (setup took ${slowSetupMs}ms before provider request started)`,
-    );
-    await vi.advanceTimersByTimeAsync(1000);
-    await assertion;
-    expect(options.signal.aborted).toBe(true);
-  });
-
-  it("rejects when image runtime setup exceeds the request timeout", async () => {
-    vi.useFakeTimers();
-    resolveModelAsyncMock.mockImplementationOnce(() => new Promise(() => {}));
-
-    const result = describeImageWithModel({
-      cfg: {},
-      agentDir: "/tmp/openclaw-agent",
-      provider: "openai",
-      model: "gpt-5.4-mini",
-      buffer: Buffer.from("png-bytes"),
-      fileName: "image.png",
-      mime: "image/png",
-      prompt: "Describe the image.",
-      timeoutMs: 25,
-    });
-
-    const assertion = expect(result).rejects.toThrow(
-      "image description setup timed out after 25ms before provider request started",
-    );
-    await vi.advanceTimersByTimeAsync(25);
-    await assertion;
-    expect(completeMock).not.toHaveBeenCalled();
-  });
-
-  it("releases a prepared generation that resolves after the setup timeout", async () => {
-    vi.useFakeTimers();
-    let finishResolution!: (value: {
-      authStorage: typeof preparedAuthStorage;
-      model: { provider: string; id: string; api: string; input: string[] };
-      modelRegistry: object;
-    }) => void;
-    resolveModelAsyncMock.mockImplementationOnce(
-      () =>
-        new Promise((resolve) => {
-          finishResolution = resolve;
-        }),
-    );
-
-    const result = describeImageWithModel({
-      cfg: {},
-      agentDir: "/tmp/openclaw-agent",
-      provider: "openai",
-      model: "gpt-5.4-mini",
-      buffer: Buffer.from("png-bytes"),
-      fileName: "image.png",
-      mime: "image/png",
-      prompt: "Describe the image.",
-      timeoutMs: 25,
-    });
-    const assertion = expect(result).rejects.toThrow(
-      "image description setup timed out after 25ms before provider request started",
-    );
-    await vi.advanceTimersByTimeAsync(25);
-    await assertion;
-
-    finishResolution({
-      authStorage: preparedAuthStorage,
-      model: {
-        provider: "openai",
-        id: "gpt-5.4-mini",
-        api: "openai-responses",
-        input: ["text", "image"],
-      },
-      modelRegistry: {},
-    });
-    await vi.runAllTimersAsync();
-    await vi.waitFor(() => expect(releasePreparedModelRuntimeMock).toHaveBeenCalledOnce());
-    expect(completeMock).not.toHaveBeenCalled();
-  });
-
-  it("normalizes deprecated google flash ids and keeps profile model/auth selection", async () => {
-    const findMock = vi.fn((provider: string, modelId: string) => {
-      expect(provider).toBe("google");
-      expect(modelId).toBe("gemini-3-flash-preview");
-      return {
-        provider: "google",
-        id: "gemini-3-flash-preview",
-        input: ["text", "image"],
-        baseUrl: "https://generativelanguage.googleapis.com/v1beta",
-      };
-    });
-    discoverModelsMock.mockReturnValue({ find: findMock });
-    completeMock.mockResolvedValue({
-      role: "assistant",
-      api: "google-generative-ai",
-      provider: "google",
-      model: "gemini-3-flash-preview",
-      stopReason: "stop",
-      timestamp: Date.now(),
-      content: [{ type: "text", text: "flash ok" }],
-    });
-
-    const result = await describeImageWithModel({
-      cfg: {},
-      agentDir: "/tmp/openclaw-agent",
-      provider: "google",
-      model: "gemini-3.1-flash-preview",
-      profile: "google:default",
-      preferredProfile: "google:preferred",
-      buffer: Buffer.from("png-bytes"),
-      fileName: "image.png",
-      mime: "image/png",
-      prompt: "Describe the image.",
-      timeoutMs: 1000,
-    });
-
-    expect(result).toEqual({
-      text: "flash ok",
-      model: "gemini-3-flash-preview",
-    });
-    expect(findMock).toHaveBeenCalled();
-    for (const call of resolveModelAsyncMock.mock.calls) {
-      expect(call[4]).toEqual(
-        expect.objectContaining({
-          authProfileId: "google:default",
-          preferredProfile: "google:preferred",
-        }),
-      );
-    }
-    const authRequest = getApiKeyForModelCall();
-    expect(authRequest?.profileId).toBe("google:default");
-    expect(authRequest?.preferredProfile).toBe("google:preferred");
-    expect(setRuntimeApiKeyMock).toHaveBeenCalledWith("google", "oauth-test");
-  });
-
-  it("keeps stable GA gemini 3.1 flash-lite ids during lookup and keeps profile auth selection", async () => {
-    const findMock = vi.fn((provider: string, modelId: string) => {
-      expect(provider).toBe("google");
-      expect(modelId).toBe("gemini-3.1-flash-lite");
-      return {
-        provider: "google",
-        id: "gemini-3.1-flash-lite",
-        input: ["text", "image"],
-        baseUrl: "https://generativelanguage.googleapis.com/v1beta",
-      };
-    });
-    discoverModelsMock.mockReturnValue({ find: findMock });
-    completeMock.mockResolvedValue({
-      role: "assistant",
-      api: "google-generative-ai",
-      provider: "google",
-      model: "gemini-3.1-flash-lite",
-      stopReason: "stop",
-      timestamp: Date.now(),
-      content: [{ type: "text", text: "flash lite ok" }],
-    });
-
-    const result = await describeImageWithModel({
-      cfg: {},
-      agentDir: "/tmp/openclaw-agent",
-      provider: "google",
-      model: "gemini-3.1-flash-lite",
-      profile: "google:default",
-      buffer: Buffer.from("png-bytes"),
-      fileName: "image.png",
-      mime: "image/png",
-      prompt: "Describe the image.",
-      timeoutMs: 1000,
-    });
-
-    expect(result).toEqual({
-      text: "flash lite ok",
-      model: "gemini-3.1-flash-lite",
-    });
-    expect(findMock).toHaveBeenCalled();
-    const authRequest = getApiKeyForModelCall();
-    expect(authRequest?.profileId).toBe("google:default");
-    expect(setRuntimeApiKeyMock).toHaveBeenCalledWith("google", "oauth-test");
-  });
-
-  it("rematerializes profile-scoped image metadata after auth selects a backup profile", async () => {
-    const authStorage = { setRuntimeApiKey: setRuntimeApiKeyMock };
-    const modelRegistry = {};
-    const hintedModel = {
-      provider: "github-copilot",
-      id: "gpt-5.6-sol",
-      api: "openai-responses",
-      input: ["text", "image"],
-      contextWindow: 200_000,
-      maxTokens: 64_000,
-    };
-    const authoritativeModel = {
-      ...hintedModel,
-      contextWindow: 1_050_000,
-      maxTokens: 128_000,
-    };
-    resolveModelAsyncMock
-      .mockResolvedValueOnce({ model: hintedModel, authStorage, modelRegistry })
-      .mockResolvedValueOnce({ model: hintedModel, authStorage, modelRegistry })
-      .mockResolvedValueOnce({ model: authoritativeModel, authStorage, modelRegistry });
-    getApiKeyForModelMock.mockResolvedValueOnce({
-      apiKey: "backup-profile-token",
-      source: "profile:github-copilot:backup",
-      mode: "token",
-      profileId: "github-copilot:backup",
-    });
-    shouldPreferProviderRuntimeResolvedModelMock.mockReturnValueOnce(true);
-    completeMock.mockResolvedValue({
-      role: "assistant",
-      api: "openai-responses",
-      provider: "github-copilot",
-      model: "gpt-5.6-sol",
-      stopReason: "stop",
-      timestamp: Date.now(),
-      content: [{ type: "text", text: "profile-scoped image ok" }],
-    });
-
-    await describeImageWithModel({
-      cfg: {},
-      agentDir: "/tmp/openclaw-agent",
-      provider: "github-copilot",
-      model: "gpt-5.6-sol",
-      profile: "github-copilot:preferred",
-      buffer: Buffer.from("png-bytes"),
-      fileName: "image.png",
-      mime: "image/png",
-      prompt: "Describe the image.",
-      timeoutMs: 1000,
-    });
-
-    expect(resolveModelAsyncMock).toHaveBeenCalledTimes(3);
-    expect(resolveModelAsyncMock.mock.calls[2]?.[4]).toEqual(
-      expect.objectContaining({
-        authStorage,
-        modelRegistry: expect.objectContaining({
-          llmRuntime: expect.anything(),
-        }),
-        authProfileId: "github-copilot:backup",
-      }),
-    );
-    const [completionModel] = requireFirstMockCall(completeMock, "complete");
-    expect(completionModel).toEqual(
-      expect.objectContaining({
-        contextWindow: 1_050_000,
-        maxTokens: 128_000,
-      }),
-    );
-  });
-
-  it("places image prompt in user content for github-copilot provider", async () => {
-    const providerStreamResult = {
-      role: "assistant",
-      api: "openai-completions",
-      provider: "github-copilot",
-      model: "gemini-3.1-pro-preview",
-      stopReason: "stop",
-      timestamp: Date.now(),
-      content: [{ type: "text", text: "A solid red square." }],
-    };
-    const providerStreamFn = vi.fn((_model: unknown, _context: unknown, _options: unknown) => ({
-      result: vi.fn(async () => providerStreamResult),
-    }));
-    registerProviderStreamForModelMock.mockReturnValueOnce(providerStreamFn);
-    discoverModelsMock.mockReturnValue({
-      find: vi.fn(() => ({
-        provider: "github-copilot",
-        id: "gemini-3.1-pro-preview",
-        input: ["text", "image"],
-        api: "openai-completions",
-        baseUrl: "https://stale.example.test",
-      })),
-    });
-
-    await describeImageWithModel({
-      cfg: {},
-      agentDir: "/tmp/openclaw-agent",
-      provider: "github-copilot",
-      model: "gemini-3.1-pro-preview",
-      buffer: Buffer.from("png-bytes"),
-      fileName: "image.png",
-      mime: "image/png",
-      prompt: "Describe the image.",
-      timeoutMs: 1000,
-    });
-
-    expect(completeMock).not.toHaveBeenCalled();
-    expect(providerStreamFn).toHaveBeenCalledOnce();
-    expect(prepareProviderRuntimeAuthMock).toHaveBeenCalledWith(
-      expect.objectContaining({
-        provider: "github-copilot",
-        context: expect.objectContaining({ apiKey: "oauth-test", authMode: "oauth" }),
-      }),
-    );
-    const storedToken = setRuntimeApiKeyMock.mock.calls[0]?.[1] as string;
-    expect(setRuntimeApiKeyMock.mock.calls[0]?.[0]).toBe("github-copilot");
-    expect(looksLikeSecretSentinel(storedToken)).toBe(true);
-    expect(storedToken).not.toBe("copilot-api-token");
-    expect(resolveSecretSentinel(storedToken)).toBe("copilot-api-token");
-    const [completionModel, context, options] = providerStreamFn.mock.calls[0] as unknown as [
-      { baseUrl?: string },
-      { systemPrompt?: string; messages?: Array<{ role: string; content: unknown[] }> },
-      { apiKey?: string; headers?: Record<string, string> },
-    ];
-    expect(completionModel.baseUrl).toBe("https://api.githubcopilot.com");
-    expect(options.apiKey).toBe(storedToken);
-    expect(options.headers).toMatchObject({
-      "Copilot-Integration-Id": "vscode-chat",
-      "Copilot-Vision-Request": "true",
-      "Editor-Version": "vscode/1.107.0",
-      "User-Agent": "GitHubCopilotChat/0.35.0",
-    });
-    expect(context.systemPrompt).toBeUndefined();
-    const userMessage = context.messages?.find((m) => m.role === "user");
-    expect(userMessage).toBeDefined();
-    const contentTypes = userMessage!.content.map((block) => (block as { type: string }).type);
-    expect(contentTypes).toContain("text");
-    expect(contentTypes).toContain("image");
-  });
-
-  it("keeps an exchanged Copilot image token opaque for sentinel-backed auth", async () => {
-    const sourceSecret = "copilot-image-source-secret";
-    const sourceSentinel = mintSecretSentinel(sourceSecret, {
-      label: "model-auth:github-copilot",
-    });
-    getApiKeyForModelMock.mockResolvedValueOnce({
-      apiKey: sourceSentinel,
-      source: "test",
-      mode: "token",
-    });
-    unwrapSecretSentinelsForProviderEgressMock.mockReturnValueOnce(sourceSecret);
-    const providerStreamFn = vi.fn((_model: unknown, _context: unknown, _options: unknown) => ({
-      result: vi.fn(async () => ({
-        role: "assistant",
-        api: "openai-completions",
-        provider: "github-copilot",
-        model: "gpt-4.1",
-        stopReason: "stop",
-        timestamp: Date.now(),
-        content: [{ type: "text", text: "ok" }],
-      })),
-    }));
-    registerProviderStreamForModelMock.mockReturnValueOnce(providerStreamFn);
-    discoverModelsMock.mockReturnValue({
-      find: vi.fn(() => ({
-        provider: "github-copilot",
-        id: "gpt-4.1",
-        input: ["text", "image"],
-        api: "openai-completions",
-      })),
-    });
-
-    await describeImageWithModel({
-      cfg: {},
-      agentDir: "/tmp/openclaw-agent",
-      provider: "github-copilot",
-      model: "gpt-4.1",
-      buffer: Buffer.from("png-bytes"),
-      fileName: "image.png",
-      mime: "image/png",
-      timeoutMs: 1000,
-    });
-
-    expect(prepareProviderRuntimeAuthMock).toHaveBeenCalledWith(
-      expect.objectContaining({
-        provider: "github-copilot",
-        context: expect.objectContaining({ apiKey: sourceSentinel, authMode: "token" }),
-      }),
-    );
-    const storedToken = setRuntimeApiKeyMock.mock.calls[0]?.[1] as string;
-    expect(looksLikeSecretSentinel(storedToken)).toBe(true);
-    expect(resolveSecretSentinel(storedToken)).toBe("copilot-api-token");
-    const streamOptions = providerStreamFn.mock.calls[0]?.[2] as { apiKey?: string };
-    expect(streamOptions.apiKey).toBe(storedToken);
-  });
-
-  it("fails github-copilot image runtime setup when token exchange fails", async () => {
-    discoverModelsMock.mockReturnValue({
-      find: vi.fn(() => ({
-        provider: "github-copilot",
-        id: "gemini-3.1-pro-preview",
-        input: ["text", "image"],
-        api: "openai-completions",
-        baseUrl: "https://api.githubcopilot.com",
-      })),
-    });
-    prepareProviderRuntimeAuthMock.mockRejectedValueOnce(
-      new Error("Copilot token exchange failed: HTTP 401"),
-    );
-
-    await expect(
-      describeImageWithModel({
-        cfg: {},
-        agentDir: "/tmp/openclaw-agent",
-        provider: "github-copilot",
-        model: "gemini-3.1-pro-preview",
-        buffer: Buffer.from("png-bytes"),
-        fileName: "image.png",
-        mime: "image/png",
-        prompt: "Describe the image.",
-        timeoutMs: 1000,
-      }),
-    ).rejects.toThrow("Copilot token exchange failed: HTTP 401");
-
-    expect(setRuntimeApiKeyMock).not.toHaveBeenCalledWith("github-copilot", "oauth-test");
-    expect(completeMock).not.toHaveBeenCalled();
-  });
-
-  it("does not place image prompt in user content for non-copilot providers", async () => {
-    discoverModelsMock.mockReturnValue({
-      find: vi.fn(() => ({
-        provider: "openai",
-        id: "gpt-4o",
-        input: ["text", "image"],
-        api: "openai-responses",
-        baseUrl: "https://api.openai.com/v1",
-      })),
-    });
-    completeMock.mockResolvedValue({
-      role: "assistant",
-      api: "openai-responses",
-      provider: "openai",
-      model: "gpt-4o",
-      stopReason: "stop",
-      timestamp: Date.now(),
-      content: [{ type: "text", text: "A solid red square." }],
-    });
-
-    await describeImageWithModel({
-      cfg: {},
-      agentDir: "/tmp/openclaw-agent",
-      provider: "openai",
-      model: "gpt-4o",
-      buffer: Buffer.from("png-bytes"),
-      fileName: "image.png",
-      mime: "image/png",
-      prompt: "Describe the image.",
-      timeoutMs: 1000,
-    });
-
-    expect(completeMock).toHaveBeenCalledOnce();
-    const [, context] = completeMock.mock.calls[0] as [
-      unknown,
-      { systemPrompt?: string; messages?: Array<{ role: string; content: unknown[] }> },
-    ];
-    // Non-Copilot providers keep prompt in system message, images in user message
-    expect(context.systemPrompt).toBe("Describe the image.");
-    const userMessage = context.messages?.find((m) => m.role === "user");
-    expect(userMessage).toBeDefined();
-    const contentTypes = userMessage!.content.map((block) => (block as { type: string }).type);
-    expect(contentTypes).not.toContain("text");
-    expect(contentTypes).toContain("image");
-  });
-
-  it("defaults image-describe maxTokens to 4096 for reasoning-capable VLMs", async () => {
-    discoverModelsMock.mockReturnValue({
-      find: vi.fn(() => ({
-        api: "openai-completions",
-        provider: "agent-plan",
-        id: "doubao-seed-2.0-pro",
-        input: ["text", "image"],
-        baseUrl: "https://ark.cn-beijing.volces.com/api/plan/v3",
-      })),
-    });
-    completeMock.mockResolvedValue({
-      role: "assistant",
-      api: "openai-completions",
-      provider: "agent-plan",
-      model: "doubao-seed-2.0-pro",
-      stopReason: "stop",
-      timestamp: Date.now(),
-      content: [{ type: "text", text: "ok" }],
-    });
-
-    await describeImageWithModel({
-      cfg: {},
-      agentDir: "/tmp/openclaw-agent",
-      provider: "agent-plan",
-      model: "doubao-seed-2.0-pro",
-      buffer: Buffer.from("png-bytes"),
-      fileName: "image.png",
-      mime: "image/png",
-      prompt: "Describe the image.",
-      timeoutMs: 1000,
-    });
-
-    const options = requireFirstMockCall(completeMock, "image completion")[2];
-    expect(options.maxTokens).toBe(4096);
-  });
-
-  it("caps image-describe maxTokens by the resolved model's own maxTokens", async () => {
-    discoverModelsMock.mockReturnValue({
-      find: vi.fn(() => ({
-        api: "openai-completions",
-        provider: "fake",
-        id: "small-vlm",
-        input: ["text", "image"],
-        baseUrl: "https://example.test",
-        maxTokens: 1024,
-      })),
-    });
-    completeMock.mockResolvedValue({
-      role: "assistant",
-      api: "openai-completions",
-      provider: "fake",
-      model: "small-vlm",
-      stopReason: "stop",
-      timestamp: Date.now(),
-      content: [{ type: "text", text: "ok" }],
-    });
-
-    await describeImageWithModel({
-      cfg: {},
-      agentDir: "/tmp/openclaw-agent",
-      provider: "fake",
-      model: "small-vlm",
-      buffer: Buffer.from("png-bytes"),
-      fileName: "image.png",
-      mime: "image/png",
-      prompt: "Describe the image.",
-      timeoutMs: 1000,
-    });
-
-    const options = requireFirstMockCall(completeMock, "image completion")[2];
-    expect(options.maxTokens).toBe(1024);
-  });
-
-  it("derives workspaceDir from agentId for image runtime resolution", async () => {
-    discoverModelsMock.mockReturnValue({
-      find: vi.fn(() => ({
-        provider: "google",
-        id: "gemini-2.5-flash",
-        api: "google-generative-ai",
-        input: ["text", "image"],
-      })),
-    });
-    completeMock.mockResolvedValue({
-      role: "assistant",
-      api: "google-generative-ai",
-      provider: "google",
-      model: "gemini-2.5-flash",
-      stopReason: "stop",
-      timestamp: Date.now(),
-      content: [{ type: "text", text: "workspace ok" }],
-    });
-    const cfg = {
-      agents: {
-        list: [
-          {
-            id: "vision-agent",
-            agentDir: "/tmp/openclaw-agent",
-            workspace: "/tmp/openclaw-workspace",
-          },
-        ],
-      },
-    };
-
-    await describeImageWithModel({
-      cfg,
-      agentId: "vision-agent",
-      agentDir: "/tmp/openclaw-agent",
-      provider: "google",
-      model: "gemini-2.5-flash",
-      buffer: Buffer.alloc(1),
-      fileName: "image.png",
-      mime: "image/png",
-      prompt: "Describe the image.",
-      timeoutMs: 1000,
-    });
-
-    expect(acquireAgentRunPreparedModelRuntimeMock).toHaveBeenCalledWith(
-      expect.objectContaining({ workspaceDir: "/tmp/openclaw-workspace" }),
-    );
-    expect(resolveModelAsyncMock).toHaveBeenCalledWith(
-      "google",
-      "gemini-2.5-flash",
-      "/tmp/openclaw-agent",
-      cfg,
-      expect.objectContaining({ workspaceDir: "/tmp/openclaw-workspace" }),
-    );
-  });
-
-  it("uses one committed prepared generation for image setup and streaming", async () => {
-    const requestedCfg: OpenClawConfig = { logging: { level: "info" } };
-    const committedCfg: OpenClawConfig = { logging: { level: "debug" } };
-    acquireAgentRunPreparedModelRuntimeMock.mockResolvedValueOnce({
-      snapshot: {
-        agentDir: "/tmp/committed-agent",
-        config: committedCfg,
-        workspaceDir: "/tmp/committed-workspace",
-        createStores: () => ({
-          authStorage: preparedAuthStorage,
-          modelRegistry: {},
-        }),
-      },
-      release: releasePreparedModelRuntimeMock,
-    });
-    discoverModelsMock.mockReturnValue({
-      find: vi.fn(() => ({
-        provider: "google",
-        id: "gemini-2.5-flash",
-        api: "google-generative-ai",
-        input: ["text", "image"],
-      })),
-    });
-    completeMock.mockResolvedValue({
-      role: "assistant",
-      api: "google-generative-ai",
-      provider: "google",
-      model: "gemini-2.5-flash",
-      stopReason: "stop",
-      timestamp: Date.now(),
-      content: [{ type: "text", text: "committed runtime" }],
-    });
-
-    await describeImageWithModel({
-      cfg: requestedCfg,
-      agentDir: "/tmp/requested-agent",
-      workspaceDir: "/tmp/requested-workspace",
-      provider: "google",
-      model: "gemini-2.5-flash",
-      buffer: Buffer.alloc(1),
-      fileName: "image.png",
-      mime: "image/png",
-      prompt: "Describe the image.",
-      timeoutMs: 1000,
-    });
-
-    expect(resolveModelAsyncMock).toHaveBeenCalledWith(
-      "google",
-      "gemini-2.5-flash",
-      "/tmp/committed-agent",
-      committedCfg,
-      expect.objectContaining({ workspaceDir: "/tmp/committed-workspace" }),
-    );
-    expect(registerProviderStreamForModelMock).toHaveBeenCalledWith({
-      model: expect.objectContaining({ id: "gemini-2.5-flash" }),
-      cfg: committedCfg,
-      agentDir: "/tmp/committed-agent",
-      workspaceDir: "/tmp/committed-workspace",
-    });
-  });
-
-  it("reuses a parent run generation without acquiring another image lease", async () => {
-    const cfg: OpenClawConfig = { logging: { level: "info" } };
-    discoverModelsMock.mockReturnValue({
-      find: vi.fn(() => ({
-        provider: "google",
-        id: "gemini-2.5-flash",
-        api: "google-generative-ai",
-        input: ["text", "image"],
-      })),
-    });
-    completeMock.mockResolvedValue({
-      role: "assistant",
-      api: "google-generative-ai",
-      provider: "google",
-      model: "gemini-2.5-flash",
-      stopReason: "stop",
-      timestamp: Date.now(),
-      content: [{ type: "text", text: "parent runtime" }],
-    });
-    const preparedModelRuntime = {
-      agentDir: "/tmp/parent-agent",
-      config: cfg,
-      workspaceDir: "/tmp/parent-workspace",
-      createStores: () => ({ authStorage: preparedAuthStorage, modelRegistry: {} }),
-    } as never;
-
-    const result = await describeImageWithModel({
-      cfg,
-      agentDir: "/tmp/parent-agent",
-      workspaceDir: "/tmp/parent-workspace",
-      preparedModelRuntime,
-      provider: "google",
-      model: "gemini-2.5-flash",
-      buffer: Buffer.alloc(1),
-      fileName: "image.png",
-      mime: "image/png",
-      prompt: "Describe the image.",
-      timeoutMs: 1000,
-    });
-
-    expect(result.text).toBe("parent runtime");
-    expect(acquireAgentRunPreparedModelRuntimeMock).not.toHaveBeenCalled();
-    expect(releasePreparedModelRuntimeMock).not.toHaveBeenCalled();
   });
 });
